+++ src/main/java/com/example/demo/domain/Users.java
@@ -13,7 +13,7 @@ public class Users {
     @Id
     @Column(name="user_id", nullable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int u_id;
+    private Long u_id;
 
     @Column(name = "AFM", nullable = false)
     private Long AFM;
@@ -44,7 +44,7 @@ public class Users {
     private List<Repairs> repairs;
 
 
-    public Users(int u_id, Long AFM, String firstName, String lastName, String email, String password, UserTypeEnum uType) {
+    public Users(Long u_id, Long AFM, String firstName, String lastName, String email, String password, UserTypeEnum uType) {
         this.u_id = u_id;
         this.AFM = AFM;
         this.firstName = firstName;
@@ -57,11 +57,11 @@ public class Users {
     private Users() {
     }
 
-    public int getU_id() {
+    public Long getU_id() {
         return u_id;
     }
 
-    public void setU_id(int u_id) {
+    public void setU_id(Long u_id) {
         this.u_id = u_id;
     }
